refactor(tc): store diagnostics in annotated modules
