+++ tooling/aihc-dev/app/tc-stackage-progress/TcStackageProgress.hs
@@ -16,7 +16,7 @@ where
 import Aihc.Hackage.Stackage (loadStackageSnapshot)
 import Aihc.Hackage.Types (PackageSpec (..))
 import Aihc.Resolve (ModuleExports, ResolveResult (..), extractInterface, resolveWithDeps)
-import Aihc.Tc (TcModuleResult (..), typecheck)
+import Aihc.Tc (TcModuleResult (..), tcModuleDiagnostics, typecheck)
 import BootInterface (bootPackageNames, loadBootInterfaces)
 import Control.Concurrent.Async (replicateConcurrently_)
 import Control.Concurrent.Chan (newChan, readChan, writeChan)
@@ -186,7 +186,7 @@ typecheckOnePackageOrThrow _offline _pkg info depExports = do
 typecheckResolvedPackage :: ResolveResult -> IO RSP.PackageStatus
 typecheckResolvedPackage resolveResult = do
   let results = typecheck (resolvedModules resolveResult)
-      !diagCount = sum (map (length . tcmDiagnostics) results)
+      !diagCount = sum (map (length . tcModuleDiagnostics . tcmModule) results)
   _ <- evaluate diagCount
   if all tcmSuccess results
     then pure (RSP.PkgSuccess (extractInterface resolveResult))
@@ -194,7 +194,7 @@ typecheckResolvedPackage resolveResult = do
 
 renderTcDiagnostics :: [TcModuleResult] -> String
 renderTcDiagnostics results =
-  unlines [show diag | result <- results, diag <- tcmDiagnostics result]
+  unlines [show diag | result <- results, diag <- tcModuleDiagnostics (tcmModule result)]
 
 reportResults :: Int -> Map Text RSP.PackageStatus -> IO ()
 reportResults topN results = do
